cpu/kinetis_common: Cast device index enums to unsigned int before comparing for out of range values
Suppresses false warning in Clang:

    warning: comparison of constant 42 with expression of type 'xxx_t'
    is always false [-Wtautological-constant-out-of-range-compare]

See:
 - https://llvm.org/bugs/show_bug.cgi?id=16154
 - https://llvm.org/bugs/show_bug.cgi?id=22062

+++ cpu/kinetis_common/spi.c
@@ -852,7 +852,7 @@ int spi_init_slave(spi_t dev, spi_conf_t conf, char(*cb)(char data))
 
 int spi_acquire(spi_t dev)
 {
-    if (dev >= SPI_NUMOF) {
+    if ((unsigned int)dev >= SPI_NUMOF) {
         return -1;
     }
 
@@ -862,7 +862,7 @@ int spi_acquire(spi_t dev)
 
 int spi_release(spi_t dev)
 {
-    if (dev >= SPI_NUMOF) {
+    if ((unsigned int)dev >= SPI_NUMOF) {
         return -1;
     }
